reorganize some things
