+++ src/reducers/actionCreators.js
@@ -1,0 +1,8 @@
+import actions from "./actions";
+
+export default {
+  toggleRules: () => ({ type: actions.TOGGLE_RULES }),
+  randomizeRealms: () => ({
+    type: actions.RANDOMIZE_REALMS,
+  }),
+};
