+++ src/reducers/actions.js
@@ -1,0 +1,4 @@
+export default {
+  TOGGLE_RULES: "rolling-realms/app/toggle-rules",
+  RANDOMIZE_REALMS: "rolling-realms/app/randomize-realms",
+};
