+++ src/reducers/app.js
@@ -1,25 +1,12 @@
-import { getInitialRealms, randomizeRealms } from "./selectors";
+import getInitialRealms from "../utilities/getInitialRealms";
+import randomizeRealms from "../utilities/randomizeRealms";
+import actions from "./actions";
 
 const initialAppState = {
   areRulesOpen: false,
   rounds: getInitialRealms(),
 };
 
-// Actions
-const actions = {
-  TOGGLE_RULES: "rolling-realms/app/toggle-rules",
-  RANDOMIZE_REALMS: "rolling-realms/app/randomize-realms",
-};
-
-// Action Creators
-export const actionCreators = {
-  toggleRules: () => ({ type: actions.TOGGLE_RULES }),
-  randomizeRealms: () => ({
-    type: actions.RANDOMIZE_REALMS,
-  }),
-};
-
-// Reducer
 const app = (state = initialAppState, action = {}) => {
   switch (action?.type) {
     case actions.TOGGLE_RULES:
