+++ src/reducers/index.js
@@ -7,6 +7,10 @@ export default () => {
     combineReducers({
       app,
     }),
+    // eslint-disable-next-line no-underscore-dangle
+    window.__REDUX_DEVTOOLS_EXTENSION__ &&
+      // eslint-disable-next-line no-underscore-dangle
+      window.__REDUX_DEVTOOLS_EXTENSION__(),
   );
 
   return store;
