+++ src/utilities/randomizeRealms.js
@@ -1,0 +1,6 @@
+import { shuffle } from "lodash";
+import getInitialRealms from "./getInitialRealms";
+
+export default (state) => {
+  return { ...state, rounds: shuffle(getInitialRealms()) };
+};
